feat: accept markdown output
Phi3 sticks to outputting pure code most of the time, but still wraps it in
backtics to form a markdown code block. The simple loop for this MVP
deletes the first and last line of each output to accustom this. To keep
the logic in the loop simple, we add a sacrifical line to the start and
end of the input file.